requirements.txt
tensorflow
gradio
opencv-python
pillow
numpy